Fixed issues with MB vs KB in config file, also added service extension so, at log2ram exit, I rerun the log purge commands so at the next boot we will not be oversided.
The log purge commands attack the largest logger first. Also, remember logs are compressed so you might end up with many GB of logs at runtime that would be committed to disk at shut down and then at next start, log2ram would fail to copy the contents into what it thinks is too small of a destination. So, if you have one very chatty app, it's logs will be the first to be attacked to reduce their size, then the next, and the next. System logs are limited to 32MB, so that's all they can do.